third_party/WebKit/LayoutTests/netinfo/type-change-no-listener.html

<!DOCTYPE html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/netinfo_common.js"></script>
</head>
<body>
<script>
description('Tests that the type changes without an active listener');

internals.setNetworkConnectionInfo(newConnectionType);
shouldBe('connection.type', 'newConnectionType');

finishJSTest();
</script>
</body>
</html>
