third_party/WebKit/LayoutTests/netinfo/type-change-no-listener-expected.txt

Tests that the type changes without an active listener

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS connection.type is newConnectionType
PASS successfullyParsed is true

TEST COMPLETE
